[algo] oom_sort: Fix access to back of FlushQueue
If data is merged in multiple rounds with multiple threads then
we get an out of bounds read after the first set of blocks has been merged.
The FlushQueue::size() function returns the size of the already flushed data+the size of the flush buffer.
This results in an access to flushQueue.back() since flushQueue.size() is not zero.